:art: Reuse panel sorting function in commands
